fixed gc overhead issue in downaloder and improved workflow
Former-commit-id: 06c0322165c853d5a44d8eaec34fc512806c12b8

+++ backend/services/downloader/src/main/scala/edu/ucdavis/fiehnlab/mona/backend/services/downloader/types/DownloaderTypes.scala
@@ -19,7 +19,7 @@ case class QueryExport(
                         emailAddress: String,
 
                         date: Date,
-                        count: Integer,
+                        count: Long,
                         size: Long,
 
                         queryFile: String,
@@ -31,8 +31,10 @@ case class PredefinedQuery(
                             @(Id@field)
                             label: String,
                             description: String,
+
                             query: String,
-                            queryCount: Integer,
+                            queryCount: Long,
+
                             jsonExport: QueryExport,
                             mspExport: QueryExport
                           )
